fix: remove test fixture fit_result
Writing with a callback in a fixture and then loading it in a test led
to instable CI. This should fix that
